Move Player to a separate file
