+++ beers/models/__init__.py
@@ -1,7 +1,9 @@
 from beers.models.models import (
     Beer, Brewery, Contest, Contest_Beer, Contest_Bonus, Contest_Brewery,
-    Contest_Checkin, Contest_Player, Player, Unvalidated_Checkin
+    Contest_Checkin, Contest_Player, Unvalidated_Checkin
 )
+
+from beers.models.player import Player
 
 __all__ = [
     'Beer', 'Brewery', 'Contest', 'Contest_Beer', 'Contest_Bonus', 'Contest_Brewery',
